ref: move state to `WizardZoneObject`
